pyproject.toml
[project]
name = "mail2whatsapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.19.0",
    "python-dotenv>=1.0.0",
    "openai-agents>=0.4.2",
    "google-api-python-client>=2.185.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
]
